pyproject.toml
[project]
name = "contextual-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3>=1.40.11",
    "crewai>=0.159.0",
    "crewai-tools>=0.62.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "google-genai>=1.30.0",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.30",
    "langchain-community>=0.3.27",
    "langchain-docling>=1.0.0",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.6",
    "langchain-text-splitters>=0.3.9",
    "llama-index>=0.13.2",
    "llama-index-embeddings-ollama>=0.8.1",
    "llama-index-llms-ollama>=0.7.1",
    "llama-index-vector-stores-postgres>=0.6.3",
    "psycopg2-binary>=2.9.10",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "trafilatura>=2.0.0",
    "uvicorn>=0.35.0",
]
